requirements.txt
Authlib==0.14.1
boto3==1.24.19
botocore==1.27.59
Flask==2.0.3
Flask_Caching==2.0.1
Flask_Cors==3.0.10
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
jsonschema==3.2.0
psycopg2_binary==2.9.3
PyMySQL==1.0.2
requests==2.28.1
shortuuid==1.0.11
SQLAlchemy==1.4.45
sqlalchemy_json==0.5.0
Werkzeug==2.0.3
flask-swagger-ui==4.11.1
apscheduler==3.10.4